+++ xarray/core/dataarray.py
@@ -1384,9 +1384,8 @@ class DataArray(AbstractArray, BaseDataObject):
             if isinstance(other, (Dataset, groupby.GroupBy)):
                 return NotImplemented
             if hasattr(other, 'indexes'):
-                # if user does not specify join, default to OPTIONS['join']
-                how_to_join = join or OPTIONS['join']
-                self, other = align(self, other, join=how_to_join, copy=False)
+                align_type = OPTIONS['arithmetic_join'] if join is None else join
+                self, other = align(self, other, join=align_type, copy=False)
             other_variable = getattr(other, 'variable', other)
             other_coords = getattr(other, 'coords', None)